requirements.txt
# Основные зависимости
ccxt==4.5.5
pandas==2.1.0
numpy==1.24.0
python-dotenv==1.0.0

# Machine Learning
torch==2.5.0
stable-baselines3==2.3.0
gymnasium==0.29.0
tensorboard==2.14.0
